+++ examples/input/storage.wgsl
@@ -1,0 +1,3 @@
+#import ~.main
+
+var<storage, read_write> rectangles: Rectangles;
